WIP: Model trait implemented for ODE and Analytical
